ref(ui) Consolidate duplicate UI components
Consolidate a few repeated UI components into a shared components. While
not all the components are always used in charts, they often are so
that's where I put them.
