+++ src/sentry/static/sentry/app/views/events/yAxisSelector.jsx
@@ -6,8 +6,7 @@ import DropdownButton from 'app/components/dropdownButton';
 import DropdownControl, {DropdownItem} from 'app/components/dropdownControl';
 import {t} from 'app/locale';
 import space from 'app/styles/space';
-
-import {InlineContainer, SectionHeading} from '../eventsV2/styles';
+import {InlineContainer, SectionHeading} from 'app/components/charts/styles';
 
 const YAxisSelector = props => {
   const {options, onChange, selected} = props;
