+++ src/sentry/static/sentry/app/views/eventsV2/styles.tsx
@@ -2,27 +2,6 @@ import styled from '@emotion/styled';
 
 import space from 'app/styles/space';
 
-export const SectionHeading = styled('h4')`
-  color: ${p => p.theme.gray3};
-  font-size: ${p => p.theme.fontSizeMedium};
-  margin: ${space(1)} 0;
-  padding-right: ${space(1)};
-  line-height: 1.2;
-`;
-
-export const SectionValue = styled('span')`
-  color: ${p => p.theme.gray3};
-  font-size: ${p => p.theme.fontSizeMedium};
-  margin-right: ${space(1)};
-`;
-
-export const ChartControls = styled('div')`
-  display: flex;
-  justify-content: space-between;
-  padding: ${space(1)} ${space(3)};
-  border-top: 1px solid ${p => p.theme.borderLight};
-`;
-
 export const ContentBox = styled('div')`
   padding: ${space(2)} ${space(4)};
   margin: 0;
@@ -53,34 +32,4 @@ export const HeaderControls = styled('div')`
   justify-self: end;
   grid-row: 1/2;
   grid-column: 2/3;
-`;
-
-export const InlineContainer = styled('div')`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-`;
-
-export const SubHeading = styled('h3')`
-  font-size: ${p => p.theme.fontSizeLarge};
-  font-weight: normal;
-  color: ${p => p.theme.gray5};
-  margin: 0;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-`;
-
-export const ErrorPanel = styled('div')`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  flex: 1;
-  flex-shrink: 0;
-  overflow: hidden;
-  height: 200px;
-  position: relative;
-  border-color: transparent;
-  margin-bottom: 0;
 `;
