+++ src/sentry/static/sentry/app/views/eventsV2/tags.tsx
@@ -8,6 +8,7 @@ import {t} from 'app/locale';
 import space from 'app/styles/space';
 import {Client} from 'app/api';
 import SentryTypes from 'app/sentryTypes';
+import {SectionHeading} from 'app/components/charts/styles';
 import EmptyStateWarning from 'app/components/emptyStateWarning';
 import {IconWarning} from 'app/icons';
 import theme from 'app/utils/theme';
@@ -19,7 +20,6 @@ import {generateQueryWithTag} from 'app/utils';
 import {trackAnalyticsEvent} from 'app/utils/analytics';
 import EventView, {isAPIPayloadSimilar} from 'app/utils/discover/eventView';
 
-import {SectionHeading} from './styles';
 import {fetchTagFacets, Tag} from './utils';
 
 type Props = {
