+++ src/sentry/static/sentry/app/views/performance/charts/styles.tsx
@@ -7,20 +7,6 @@ export const HeaderTitle = styled('span')`
   padding: 0 ${space(1)};
 `;
 
-export const SectionHeading = styled('h4')`
-  color: ${p => p.theme.gray3};
-  font-size: ${p => p.theme.fontSizeMedium};
-  margin: ${space(1)} 0;
-  padding-right: ${space(1)};
-  line-height: 1.2;
-`;
-
-export const SectionValue = styled('span')`
-  color: ${p => p.theme.gray3};
-  font-size: ${p => p.theme.fontSizeMedium};
-  margin-right: ${space(1)};
-`;
-
 export const ChartsContainer = styled('div')`
   padding: ${space(2)} ${space(1.5)};
 `;
@@ -33,11 +19,4 @@ export const ChartsGrid = styled('div')`
 
 export const ChartContainer = styled('div')`
   flex-grow: 1;
-`;
-
-export const ChartControls = styled('div')`
-  display: flex;
-  align-items: center;
-  padding: ${space(1)} ${space(3)};
-  border-top: 1px solid ${p => p.theme.borderLight};
 `;
