+++ src/sentry/static/sentry/app/views/releasesV2/detail/overview/styles.tsx
@@ -1,15 +1,12 @@
 import styled from '@emotion/styled';
 
+import {SectionHeading as BaseSectionHeading} from 'app/components/charts/styles';
 import space from 'app/styles/space';
-
-export const SectionHeading = styled('h4')`
-  color: ${p => p.theme.gray3};
-  font-size: ${p => p.theme.fontSizeMedium};
-  margin: 0 0 ${space(1.5)} 0;
-  padding-right: ${space(1)};
-  line-height: 1.2;
-`;
 
 export const Wrapper = styled('div')`
   margin-bottom: ${space(4)};
+`;
+
+export const SectionHeading = styled(BaseSectionHeading)`
+  margin: 0 0 ${space(1.5)} 0;
 `;
